Change module art helpers into class
