+++ src/scripts/set-game-pf2e.ts
@@ -25,12 +25,12 @@ import {
 } from "@scripts/macros";
 import { remigrate } from "@scripts/system/remigrate";
 import { ActionMacros } from "@system/action-macros";
+import { CheckPF2e } from "@system/check";
 import { ConditionManager } from "@system/conditions";
 import { EffectTracker } from "@system/effect-tracker";
-import { CheckPF2e } from "@system/check";
+import { ModuleArt } from "@system/module-art";
 import { TextEditorPF2e } from "@system/text-editor";
 import { sluggify } from "@util";
-import { registerModuleArt } from "./register-module-art";
 
 /** Expose public game.pf2e interface */
 export const SetGamePF2e = {
@@ -65,7 +65,7 @@ export const SetGamePF2e = {
             licenseViewer: new LicenseViewer(),
             rollActionMacro,
             rollItemMacro,
-            system: { moduleArt: { map: new Map(), refresh: registerModuleArt }, remigrate, sluggify },
+            system: { moduleArt: new ModuleArt(), remigrate, sluggify },
             variantRules: { AutomaticBonusProgression },
         };
